move coverage methods to seperate class
