+++ old/globals.py
